Changed i32 to i64

+++ src/scanning_data_format.rs
@@ -20,7 +20,7 @@ pub struct ScanningData {
     pub production_rate: u32,
     pub game_over: Option<i32>,
     pub turn_based: Option<i32>,
-    pub turn_based_time_out: Option<i32>,
+    pub turn_based_time_out: Option<i64>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
